Delete deprecated files
