+++ SAPSec.Core/SAPSec.Core.csproj
@@ -7,10 +7,13 @@
   </PropertyGroup>
 
   <ItemGroup>
+    <Compile Remove="Features\Secondary\Ks4HeadlineMeasures\**" />
     <Compile Remove="Model\KS4\**" />
     <Compile Remove="Model\Search\**" />
+    <EmbeddedResource Remove="Features\Secondary\Ks4HeadlineMeasures\**" />
     <EmbeddedResource Remove="Model\KS4\**" />
     <EmbeddedResource Remove="Model\Search\**" />
+    <None Remove="Features\Secondary\Ks4HeadlineMeasures\**" />
     <None Remove="Model\KS4\**" />
     <None Remove="Model\Search\**" />
   </ItemGroup>
@@ -30,10 +33,6 @@
 
   <ItemGroup>
     <ProjectReference Include="..\Data\SAPSec.Data\SAPSec.Data.csproj" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <Folder Include="Features\Secondary\Ks4HeadlineMeasures\" />
   </ItemGroup>
 
 </Project>
